frontend/settings: split electrum.jsx into one file per component
Easier to convert to typescript piece by piece.
